refactoring
Even though i would like to kill this module asap, it will still be
staying with us for a while. This module should die because there
must be something useable on the forge and i don't want to manage
this one long-term. I still am refactoring it just a bit so i can
finaly implement the puppet infra i want as per radiorabe/puppet.